+++ lib/app.dart
@@ -1,14 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:todo/ui/counter_screen.dart';
+import 'package:todo/presentation/pages/counter_screen.dart';
 
-class MainApp extends StatefulWidget {
+class MainApp extends StatelessWidget {
   const MainApp({super.key});
 
-  @override
-  State<MainApp> createState() => _MainAppState();
-}
-
-class _MainAppState extends State<MainApp> {
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
